pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "python-multipart<1.0.0,>=0.0.7",
    "email-validator<3.0.0.0,>=2.1.0.post1",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic>2.0",
    "emails<1.0,>=0.6",
    "jinja2<4.0.0,>=3.1.4",
    "alembic<2.0.0,>=1.12.1",
    "httpx<1.0.0,>=0.25.1",
    "psycopg[binary] (>=3.2.13,<4.0.0)",
    "sqlmodel<1.0.0,>=0.0.21",
    # Pin bcrypt until passlib supports the latest
    "bcrypt==4.3.0",
    "pydantic-settings<3.0.0,>=2.2.1",
    "sentry-sdk[fastapi]<2.0.0,>=1.40.6",
    "pyjwt<3.0.0,>=2.8.0",
    "pymongo (>=4.15.4,<5.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "psutil (>=7.1.3,<8.0.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
    "aiosmtplib (>=5.0.0,<6.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg-binary (>=3.3.0,<4.0.0)",
    "starlette (==0.46.0)",
    "fastapi-cli (==0.0.7)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "limits (>=5.6.0,<6.0.0)",
    "async-timeout (>=5.0.1,<6.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "typer[all] (>=0.20.0,<0.21.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "qrcode[pil] (>=8.2,<9.0)",
    "geoip2 (>=5.2.0,<6.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "scalar-fastapi (>=1.5.0,<2.0.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true

[tool.poetry.scripts]
novakit = "cli.main:main"

